move error from dashboard to withConfig
